Implementing-a-Parallel-Gauss-Jordan-algorithm-on-GPU-using-CUDA-architecture
In this project, we have exploited the parallelizing capabilities of
GPUs and have tried comparing the GPU general purpose implementation
with that of CPU implementation. Our goal in this project is to check
the efficiency of the CUDA Toolkit so as to use it as an advantage to
parallelize algorithms for their faster execution.